Fix is_pressing input logic
Previously, the `is_pressing` function only returned `true` if the exact
keys in `io_bits` were pressed. The new logic properly checks for the
bits of the button(s), even if other buttons are pressed simultaneously.

This fix should correctly show the "A Press" state in the Non-CFW view
for Crystal.